Change uint32 ms types to double

+++ glance_registry_local_check.py
@@ -49,7 +49,7 @@ def check(auth_ref, args):
     # only want to send other metrics if api is up
     if is_up:
         milliseconds = r.elapsed.total_seconds() * 1000
-        metric('glance_registry_local_response_time', 'uint32',
+        metric('glance_registry_local_response_time', 'double',
                '%.3f' % milliseconds, 'ms')
